requirements.txt
telethon==1.33.5
flask==3.0.0
flask-cors==4.0.0
mongodb-driver==0.0.1
pymongo==4.6.0
mongoengine==0.27.0
python-dotenv==1.0.0
pytz==2023.3
jdatetime==4.0.0
requests==2.31.0
APScheduler==3.10.4
werkzeug==3.0.1
google-cloud-translate==3.11.1
python-telegram-bot==20.3
pycryptodome==3.18.0

